refactor(xapiri): extract dashboard_focus.go — visibility + focus navigation
Part 3 of ADR 0014 dashboard split. Moves isCloud, onPremProviders,
providerListForMode, rebuildProviderList, isHidden, visibleFocusList,
focusAtConfigRow, jumpFocus, and moveFocus out of dashboard.go.

Trap 2 compliance: isHidden cross-references toggles and selects across all
tabs and remains whole in dashboard_focus.go. No iota constants in any
tab_*.go file.
